Neaten up some code

+++ Simple_UEFI_Bootloader/src/Bootloader.c
@@ -396,7 +396,7 @@ EFI_STATUS efi_main(EFI_HANDLE ImageHandle, EFI_SYSTEM_TABLE *SystemTable)
 
   // Data verification
   Print(L"Config table address: 0x%llx\r\n", ST->ConfigurationTable);
-  Print(L"Data at RSDP (first 16 bytes): 0x%016llx%016llx\r\n", *(EFI_PHYSICAL_ADDRESS*)(ST->ConfigurationTable[RSDP_index].VendorTable + 8), *(EFI_PHYSICAL_ADDRESS*)ST->ConfigurationTable[RSDP_index].VendorTable);
+  Print(L"Data at RSDP (first 16 bytes): 0x%016llx%016llx\r\n", *(EFI_PHYSICAL_ADDRESS*)(((UINT64)ST->ConfigurationTable[RSDP_index].VendorTable) + 8), *(EFI_PHYSICAL_ADDRESS*)ST->ConfigurationTable[RSDP_index].VendorTable);
 #endif
 
   // Load a program and exit boot services, then pass a loader block to that program's entry point to execute the program
